--- store/user-service/src/main/java/by/innowise/user_service/UserServiceApplication.java
@@ -1,14 +1,0 @@
-package by.innowise.user_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-
-@EnableCaching
-@SpringBootApplication
-public class UserServiceApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(UserServiceApplication.class, args);
-  }
-}
