--- store/user-service/src/main/java/by/innowise/user_service/config/cache/RedisConfig.java
@@ -1,50 +1,0 @@
-package by.innowise.user_service.config.cache;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import java.time.Duration;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-@EnableCaching
-public class RedisConfig {
-
-  public static final String USER_CACHE = "users";
-  public static final String CARD_INFO_CACHE = "cards";
-
-  @Bean
-  public RedisCacheManager cacheManager(RedisConnectionFactory connectionFactory) {
-    ObjectMapper objectMapper = new ObjectMapper();
-    objectMapper.registerModule(new JavaTimeModule());
-    objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-    objectMapper.activateDefaultTyping(
-        objectMapper.getPolymorphicTypeValidator(),
-        ObjectMapper.DefaultTyping.NON_FINAL
-    );
-
-    GenericJackson2JsonRedisSerializer serializer = new GenericJackson2JsonRedisSerializer(
-        objectMapper);
-
-    RedisCacheConfiguration config = RedisCacheConfiguration.defaultCacheConfig()
-        .entryTtl(Duration.ofHours(1))
-        .serializeKeysWith(
-            RedisSerializationContext.SerializationPair.fromSerializer(new StringRedisSerializer()))
-        .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(serializer))
-        .disableCachingNullValues();
-
-    return RedisCacheManager.builder(connectionFactory)
-        .cacheDefaults(config)
-        .transactionAware()
-        .build();
-  }
-
-}
