--- store/user-service/src/main/java/by/innowise/user_service/exception/runtime/EntityNotFoundException.java
@@ -1,8 +1,0 @@
-package by.innowise.user_service.exception.runtime;
-
-public class EntityNotFoundException extends RuntimeException {
-
-  public EntityNotFoundException(String message) {
-    super(message);
-  }
-}
