--- store/user-service/src/main/java/by/innowise/user_service/model/response/ErrorResponse.java
@@ -1,17 +1,0 @@
-package by.innowise.user_service.model.response;
-
-import java.time.LocalDateTime;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ErrorResponse {
-
-  private String title;
-  private String name;
-  private int status;
-  private String message;
-  private String path;
-  private LocalDateTime timestamp;
-}
