--- store/user-service/src/main/java/by/innowise/user_service/model/response/ListResponse.java
@@ -1,16 +1,0 @@
-package by.innowise.user_service.model.response;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ListResponse<T> {
-
-  private List<T> items;
-}
