--- store/user-service/src/main/java/by/innowise/user_service/repository/CardInfoRepository.java
@@ -1,33 +1,0 @@
-package by.innowise.user_service.repository;
-
-import by.innowise.user_service.exception.runtime.EntityNotFoundException;
-import by.innowise.user_service.model.entity.CardInfo;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CardInfoRepository extends JpaRepository<CardInfo, UUID> {
-
-  Boolean existsByNumberAndDeletedFalse(String number);
-
-  List<CardInfo> findAllByIdInAndDeletedFalse(Set<UUID> ids);
-
-  Optional<CardInfo> findByIdAndDeletedFalse(UUID id);
-
-  default boolean existsByNumber(String number) {
-    return existsByNumberAndDeletedFalse(number);
-  }
-
-  default List<CardInfo> findAllByIdIn(Set<UUID> ids) {
-    return findAllByIdInAndDeletedFalse(ids);
-  }
-
-  default CardInfo findCardInfoById(UUID id) {
-    return findByIdAndDeletedFalse(id)
-        .orElseThrow(() -> new EntityNotFoundException("Card info not found with id: " + id));
-  }
-}
