--- store/user-service/src/main/java/by/innowise/user_service/service/impl/UserServiceImpl.java
@@ -1,154 +1,0 @@
-package by.innowise.user_service.service.impl;
-
-import by.innowise.user_service.config.cache.RedisConfig;
-import by.innowise.user_service.mapper.DtoMapper;
-import by.innowise.user_service.mapper.EntityMapper;
-import by.innowise.user_service.model.dto.UserDto;
-import by.innowise.user_service.model.dto.UserPatchDto;
-import by.innowise.user_service.model.entity.User;
-import by.innowise.user_service.model.response.ListResponse;
-import by.innowise.user_service.repository.UserRepository;
-import by.innowise.user_service.service.UserService;
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
-import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.CacheManager;
-import org.springframework.cache.annotation.CacheConfig;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.cache.annotation.Caching;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@RequiredArgsConstructor
-@Service
-@CacheConfig(cacheNames = RedisConfig.USER_CACHE)
-public class UserServiceImpl implements UserService {
-
-  private final DtoMapper dtoMapper;
-  private final EntityMapper entityMapper;
-  private final UserRepository repository;
-  private final CacheManager cacheManager;
-
-  @Override
-  @Transactional
-  @CachePut(key = "#result.id")
-  public UserDto createUser(UserDto userDto) {
-    User user = entityMapper.toUser(userDto);
-    User savedUser = repository.save(user);
-
-    return dtoMapper.toUserDto(savedUser);
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  @Cacheable(key = "#id")
-  public UserDto getUserById(UUID id) {
-    User user = repository.findUserById(id);
-
-    return dtoMapper.toUserDto(user);
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  @Cacheable(key = "'ids_' + #ids.hashCode()")
-  public ListResponse<UserDto> getUsersByIds(Set<UUID> ids) {
-    List<UserDto> users = repository.findAllByIdIn(ids).stream()
-        .map(dtoMapper::toUserDto)
-        .toList();
-
-    return ListResponse.<UserDto>builder()
-        .items(users)
-        .build();
-  }
-
-  @Override
-  @Transactional(readOnly = true)
-  @Cacheable(key = "#email")
-  public UserDto getUserByEmail(String email) {
-    User user = repository.getByEmail(email);
-
-    return dtoMapper.toUserDto(user);
-  }
-
-  @Override
-  @Transactional
-  @Caching(
-      put = @CachePut(key = "#id"),
-      evict = @CacheEvict(key = "#result.email", condition = "#result.email != null")
-  )
-  public UserDto updateUser(UUID id, UserDto userDto) {
-    User user = repository.findUserById(id);
-
-    String oldEmail = user.getEmail();
-
-    user.update(userDto);
-    UserDto updated = dtoMapper.toUserDto(user);
-
-    if (!oldEmail.equals(updated.getEmail())) {
-      Objects.requireNonNull(cacheManager.getCache(RedisConfig.USER_CACHE)).evict(oldEmail);
-    }
-
-    return updated;
-  }
-
-  @Override
-  @Transactional
-  @Caching(
-      put = @CachePut(key = "#id"),
-      evict = @CacheEvict(key = "#result.email", condition = "#result.email != null")
-  )
-  public UserDto patchUser(UUID id, UserPatchDto patchDto) {
-    User user = repository.findUserById(id);
-
-    String oldEmail = user.getEmail();
-
-    user.patch(patchDto);
-    UserDto updated = dtoMapper.toUserDto(user);
-
-    if (!oldEmail.equals(updated.getEmail())) {
-      Objects.requireNonNull(cacheManager.getCache(RedisConfig.USER_CACHE)).evict(oldEmail);
-    }
-
-    return updated;
-  }
-
-  @Override
-  @Transactional
-  @Caching(
-      evict = {
-          @CacheEvict(key = "#id"),
-          @CacheEvict(key = "#result.email", condition = "#result.email != null")
-      }
-  )
-  public UserDto softDeleteUser(UUID id) {
-    User user = repository.findUserById(id);
-    user.setDeleted(true);
-
-    return dtoMapper.toUserDto(user);
-  }
-
-  @Override
-  @Transactional
-  @Caching(
-      evict = {
-          @CacheEvict(key = "#id"),
-          @CacheEvict(key = "#result.email", condition = "#result.email != null")
-      }
-  )
-  public UserDto hardDeleteUser(UUID id) {
-    User user = repository.findUserById(id);
-
-    user.getCardInfos().forEach(card ->
-        Objects.requireNonNull(cacheManager.getCache(RedisConfig.CARD_INFO_CACHE))
-            .evict(card.getId())
-    );
-
-    repository.deleteById(id);
-
-    return dtoMapper.toUserDto(user);
-  }
-}
