+++ store/user-service/src/main/java/by/innowise/userservice/exception/runtime/EntityNotFoundException.java
@@ -1,0 +1,8 @@
+package by.innowise.userservice.exception.runtime;
+
+public class EntityNotFoundException extends RuntimeException {
+
+  public EntityNotFoundException(String message) {
+    super(message);
+  }
+}
