+++ store/user-service/src/main/java/by/innowise/userservice/mapper/EntityMapper.java
@@ -1,0 +1,16 @@
+package by.innowise.userservice.mapper;
+
+import by.innowise.userservice.model.dto.CardInfoDto;
+import by.innowise.userservice.model.dto.UserDto;
+import by.innowise.userservice.model.entity.CardInfo;
+import by.innowise.userservice.model.entity.User;
+import org.mapstruct.Mapper;
+import org.mapstruct.MappingConstants;
+import org.springframework.stereotype.Component;
+
+@Component
+@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
+public interface EntityMapper {
+  User toUser(UserDto userDto);
+  CardInfo toCardInfo(CardInfoDto cardInfoDto);
+}
