+++ store/user-service/src/main/java/by/innowise/userservice/model/response/ListResponse.java
@@ -1,0 +1,16 @@
+package by.innowise.userservice.model.response;
+
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ListResponse<T> {
+
+  private List<T> items;
+}
