--- store/user-service/src/test/java/by/innowise/user_service/UserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package by.innowise.user_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
