+++ store/user-service/src/test/java/by/innowise/userservice/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package by.innowise.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
